fix some compile warnings

+++ kern/mm/memlayout.h
@@ -67,7 +67,7 @@ typedef uintptr_t pde_t;
  * */
 struct Page {
     atomic_t ref;                   // page frame's reference counter
-    uint32_t flags;                 // array of flags that describe the status of the page frame
+    unsigned long flags;                 // array of flags that describe the status of the page frame
     unsigned int property;          // used in buddy system, stores the order (the X in 2^X) of the continuous memory block
     int zone_num;                   // used in buddy system, the No. of zone which the page belongs to
     list_entry_t page_link;         // free list link
